+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Tantivy 0.10.0
 
 - Added an ASCII folding filter (@drusellers)
 - Bugfix in `query.count` in presence of deletes (@pmasurel)
+- Added `.explain(...)` in `Query` and `Weight` to (@pmasurel)
 
 Minor
 ---------